requirements.txt
deepspeed==0.15.4
hydra-core==1.3.2
omegaconf==2.3.0
torch==2.5.1
tqdm==4.67.1
transformers==4.46.1
trl==0.12.1
vllm==0.6.4.post1
